requirements.txt
# data4tr - Dependencies
# Türkçe veri seti oluşturma aracı için gerekli paketler

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration
pyyaml>=6.0

# Data processing
pandas>=2.0.0

# Optional: AI integration (future)
# openai>=1.0.0
# anthropic>=0.3.0

# Development
pytest>=7.4.0
black>=23.0.0

